chang to long

+++ src/Main.java
@@ -1,8 +1,8 @@
 public class Main {
     public static void main(String[] args) {
-        int[] sales = new int[]{5, 4, 500, 900};
+        long[] sales = new long[]{5, 4, 500, 900};
         SalesManager salesManager = new SalesManager(sales);
-        int max = salesManager.max();
+        long max = salesManager.max();
         System.out.println("Наибольшая продажа: " + max);
     }
 }
